Update payment information

+++ src/api/services/PaymentInformationService.ts
@@ -4,6 +4,7 @@ import { Logger as LoggerType } from '../../core/Logger';
 import { Types, Core, Targets } from '../../constants';
 import { validate, request } from '../../core/api/Validate';
 import { NotFoundException } from '../exceptions/NotFoundException';
+import { MessageException } from '../exceptions/MessageException';
 import { PaymentInformationRepository } from '../repositories/PaymentInformationRepository';
 import { PaymentInformation } from '../models/PaymentInformation';
 import { PaymentInformationCreateRequest } from '../requests/PaymentInformationCreateRequest';
@@ -25,18 +26,8 @@ export class PaymentInformationService {
         this.log = new Logger(__filename);
     }
 
-    @validate()
-    public async rpcFindAll( @request(RpcRequest) data: any): Promise<Bookshelf.Collection<PaymentInformation>> {
-        return this.findAll();
-    }
-
     public async findAll(): Promise<Bookshelf.Collection<PaymentInformation>> {
         return this.paymentInformationRepo.findAll();
-    }
-
-    @validate()
-    public async rpcFindOne( @request(RpcRequest) data: any): Promise<PaymentInformation> {
-        return this.findOne(data.params[0]);
     }
 
     public async findOne(id: number, withRelated: boolean = true): Promise<PaymentInformation> {
@@ -48,30 +39,13 @@ export class PaymentInformationService {
         return paymentInformation;
     }
 
-    @validate()
-    public async rpcCreate( @request(RpcRequest) data: any): Promise<PaymentInformation> {
-        return this.create({
-            type: data.params[0],
-            escrow: {
-                type: data.params[1],
-                ratio: {
-                    buyer: data.params[2],
-                    seller: data.params[3]
-                }
-            },
-            itemPrice: {
-                currency: data.params[4],
-                basePrice: data.params[5],
-                shippingPrice: {
-                    domestic: data.params[6],
-                    international: data.params[7]
-                },
-                address: {
-                    type: data.params[8],
-                    address: data.params[9]
-                }
-            }
-        });
+    public async updateByListingId(id: number, @request(PaymentInformationUpdateRequest) body: any): Promise<PaymentInformation> {
+        const paymentInformation = await this.paymentInformationRepo.findOneByListingItemTemplateId(id);
+        if (paymentInformation === null) {
+            this.log.warn(`PaymentInformation with the listing_item_template_id=${id} was not found!`);
+            throw new MessageException(`PaymentInformation with the listing_item_template_id=${id} was not found!`);
+        }
+        return this.update(paymentInformation.id, body);
     }
 
     @validate()
@@ -87,8 +61,10 @@ export class PaymentInformationService {
         const paymentInformation = await this.paymentInformationRepo.create(body);
 
         // then create escrow
-        escrow.payment_information_id = paymentInformation.Id;
-        await this.escrowService.create(escrow);
+        if (escrow) {
+            escrow.payment_information_id = paymentInformation.Id;
+            await this.escrowService.create(escrow);
+        }
 
         // then create item price
         itemPrice.payment_information_id = paymentInformation.Id;
@@ -100,52 +76,26 @@ export class PaymentInformationService {
     }
 
     @validate()
-    public async rpcUpdate( @request(RpcRequest) data: any): Promise<PaymentInformation> {
-        return this.update(data.params[0], {
-            type: data.params[1],
-            escrow: {
-                type: data.params[2],
-                ratio: {
-                    buyer: data.params[3],
-                    seller: data.params[4]
-                }
-            },
-            itemPrice: {
-                currency: data.params[5],
-                basePrice: data.params[6],
-                shippingPrice: {
-                    domestic: data.params[7],
-                    international: data.params[8]
-                },
-                address: {
-                    type: data.params[9],
-                    address: data.params[10]
-                }
-            }
-        });
-    }
-
-    @validate()
     public async update(id: number, @request(PaymentInformationUpdateRequest) body: any): Promise<PaymentInformation> {
 
         // find the existing one without related
         const paymentInformation = await this.findOne(id, false);
-
         // set new values
         paymentInformation.Type = body.type;
 
         // update paymentInformation record
         const updatedPaymentInformation = await this.paymentInformationRepo.update(id, paymentInformation.toJSON());
 
-        // find related record and delete it
-        let relatedEscrow = updatedPaymentInformation.related('Escrow').toJSON();
-        await this.escrowService.destroy(relatedEscrow.id);
+        if (body.escrow) {
+            // find related record and delete it
+            let relatedEscrow = updatedPaymentInformation.related('Escrow').toJSON();
+            await this.escrowService.destroy(relatedEscrow.id);
 
-        // recreate related data
-        relatedEscrow = body.escrow;
-        relatedEscrow.payment_information_id = id;
-        await this.escrowService.create(relatedEscrow);
-
+            // recreate related data
+            relatedEscrow = body.escrow;
+            relatedEscrow.payment_information_id = id;
+            await this.escrowService.create(relatedEscrow);
+        }
         // find related record and delete it
         let relatedItemPrice = updatedPaymentInformation.related('ItemPrice').toJSON();
         await this.itemPriceService.destroy(relatedItemPrice.id);
@@ -158,12 +108,6 @@ export class PaymentInformationService {
         // finally find and return the updated paymentInformation
         const newPaymentInformation = await this.findOne(id);
         return newPaymentInformation;
-
-    }
-
-    @validate()
-    public async rpcDestroy( @request(RpcRequest) data: any): Promise<void> {
-        return this.destroy(data.params[0]);
     }
 
     public async destroy(id: number): Promise<void> {
